Updated some task tracking scripts
Added an active project marker to simplify looking up projects that I am
currently working on in Taskwarrior. Tweaked the remaining scripts to
work with this and to drop stderr when it's doing task clock and active
project checks (so I get the fail messages I want because it really does
bother me that much).