Tweaks to the active project script
By putting the directory of a repository in the task project
description, when active is invoked, it will also invoke register -r so
if I turn off my system and forget to commit any changes, it will
automatically be handled by the off script.

Also added a journal shortcut script.