Updated i3 status interpereter for generic Python3
The reason why 3.6 was specified was that, at the time the script was
written, Python 3.4 was the standard Python 3 interpereter in Fedora. I
was, however, using string interpolation that was first introduced in
Python 3.6, hence the need to specify. Now the default is 3.7, so it's
not necessary anymore (and causes problems with out of the box Fedora
29)